+++ src/core/mod.rs
@@ -10,9 +10,11 @@ mod four_char_code;
 mod object_type;
 mod os_err;
 mod os_status;
+mod os_type;
 
 pub use arc::*;
 pub use four_char_code::*;
 pub use object_type::*;
 pub use os_err::*;
 pub use os_status::*;
+pub use os_type::*;