+++ src/core/os_type.rs
@@ -1,0 +1,4 @@
+use super::FourCharCode;
+
+/// A four-character code that is often used to indicate the type of some data.
+pub type OSType = FourCharCode;